Drivers

+++ kernel/libk/aex/time.h
@@ -2,4 +2,4 @@
 
 #include "aex/time/sleep.c"
 
-void sleep(size_t delay);
+void sleep(long delay);
